UBUNTU: [Config] armel Update versatile initrd configs
Bump CONFIG_BLK_DEV_RAM_SIZE as 4096 was too small for the unpacked size
of our current d-i versatile initrd (~11 MB); 65536 is probably fine as
on the other architectures.  Also, set CONFIG_CRAMFS=y as apparently
qemu loads the -initrd arg in a way which only works with the kernel's
initrd support instead of the initramfs code; this is a workaround until
qemu's ARM loader is fixed.

BugLink: https://bugs.launchpad.net/bugs/524893

Signed-off-by: Loïc Minier <loic.minier@ubuntu.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>