UBUNTU: perf -- build the kernel carried tools
The kernel source now carries the source for tools which are very closly
tied to the kernel version, initially the perf tool.  These tools are
included because they are tightly tied to the kernel internal ABIs.

Build these tools as part of the kernel build process and package them
up with the binary kernel packages.  As we can have more than one
version of the kernel available and because you need to use the version
which matches the version of the kernel that you are booted from,
install them using the kernel version number and flavour:

	/usr/bin/perf-2.6.32-15-generic

Signed-off-by: Andy Whitcroft <apw@canonical.com>