UBUNTU: [Config] CONFIG_HID=m
There have been a number of comments about HID being built into the kernel
with regard to DKMS and other external HID modules. Since HID is likely to
have little impact on boot speed (the original incentive for building it in),
then build core HID as a module.

Some of that discussion is buried in this thread:

https://lists.ubuntu.com/archives/kernel-team/2010-March/009202.html

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>